Debugging
Problem: Only the cluster IDs that are present in the first sample are considered when analysing other samples. For example, if no cells belonging to 'cluster 4' were present in sample 1, then cluster 4 was not interrogated in any other sample. 

Solution: Modified the script to read all the possible cluster IDs from all samples using rbindlist.